Run clippy on guests in CI (#237)
* Clippy guests


* Add matrix target to recipe


* fix broken rebase


---------

+++ src/hyperlight_guest/src/guest_function_call.rs
@@ -58,7 +58,7 @@ pub(crate) fn call_guest_function(function_call: FunctionCall) -> Result<Vec<u8>
 
         let p_function = unsafe {
             let function_pointer = registered_function_definition.function_pointer;
-            core::mem::transmute::<i64, GuestFunc>(function_pointer)
+            core::mem::transmute::<usize, GuestFunc>(function_pointer)
         };
 
         p_function(&function_call)
